+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to the [Haskell Package Versioning Policy](https://pvp.
 
 ### Changed
 - Support GHC 8.8, GHC 8.10
+- Depend on `termbox ==0.3`
+
+### Removed
+- Removed `Scene`
 
 ## [0.2.0] - 2019-06-28